Proofreading
line 23: "end-to-end" as an adjective is hardly _gradable_, I tried to make it sound less _doge_ by dropping the "very".
line 40: Speculos is an emulator as it relies on CPU emulation (from Qemu) to run native BOLOS apps built for the Nano architecture. A simulator would recreate the behavior of the apps with binaries native to the host (x86-64).